Asynchronous loading: images work.
Now images are loaded asynchronously. A full re-layout of the page
happens if any of the <img> did not have specified dimensions.

`domHTMLContent` was too pure to be true, it was refactored into
`domEndHTMLElement`

I tried making <img> and <input> containers, with their contents in
the child node; `fixupContent` was an easier solution that did not
complicate other code.