Images from `data:` URLs are back with an ugly hack.
This was surprisingly complicated: "short-circuiting" resource loading
is not easy with the new asyncrhonous model with several layers. I also
had to get rid of domHTMLContent and move the decision logic to
domEndHTMLElement.

I want to keep the DocumentT monad free of assumptions about its
underlying monad, that's why it's not possible to use SDL to decode
images in it. ImageContent needs to have a valid URL when creted, and
the network-uri library does not handle all `data:`-urls well. I tried
changing ImageContent to have `Either ? URL`, but the question mark did
not resolve itself into anything sensible.

The current hack is: domEndHTMLElement handles an <img>, looks up its
`src`. If the `src` of a node `nid` starts with `data:`, a resource with
URI like "blob:data@42" is requested. handlePageStreaming assumes that
the resource content comes from the data-encoded `src` attribute of the
node and short-circuits loading of this resource.