domMatchSelector, rulechainsForMatcher, styleFromRulechain
`domMatchSelector :: ElementRef -> Selector -> Maybe Specificity`
is used to match an element on a selector. If matched, Specificity is
computed, otherwise it's Nothing.

```
rulechainsForMatcher :: CSSRules -> TagAttrs
                     -> (Selector -> Maybe Specifity)
                     -> ([RuleID], [RuleID])
```
usually takes a `domMatchSelector eltref` and produces two rulechains:
one for important styles and one for regular. Rulechains are sorted by
specificity in descending order.